requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.12.2
cryptography==41.0.7
ShopifyAPI==12.5.0